requirements.txt
requests
psycopg2-binary
lxml
black